requirements.txt
flask==3.1.*
anthropic>=0.40.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.150.0
openpyxl>=3.1.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
